GiverController delete, WishlistController is doing the job
